bugref:9416. changed uint32_t to ULONG.
svn:sync-xref-src-repo-rev: r130428

+++ src/VBox/Main/src-server/ApplianceImplImport.cpp
@@ -1,4 +1,4 @@
-/* $Id: ApplianceImplImport.cpp 78428 2019-05-07 11:03:49Z valery.portnyagin@oracle.com $ */
+/* $Id: ApplianceImplImport.cpp 78429 2019-05-07 11:12:08Z valery.portnyagin@oracle.com $ */
 /** @file
  * IAppliance and IVirtualSystem COM class implementations.
  */
@@ -1230,8 +1230,8 @@ HRESULT Appliance::i_gettingCloudData(TaskCloud *pTask)
 
         m->virtualSystemDescriptions.clear();//clear all for assurance before creating new
         std::vector<ComPtr<IVirtualSystemDescription> > vsdArray;
-        uint32_t requestedVSDnums = 1;
-        uint32_t newVSDnums = 0;
+        ULONG requestedVSDnums = 1;
+        ULONG newVSDnums = 0;
         hrc = createVirtualSystemDescriptions(requestedVSDnums, &newVSDnums);
         if (FAILED(hrc)) throw hrc;
         if (requestedVSDnums != newVSDnums)
